requirements.txt
pillow
aiohttp
pyrogram
apscheduler
